# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Remake of my previous attampted on creating API with Node.js",
        "title": "Gin API",
        "contact": {
            "name": "API Support",
            "url": "https://mortis-grimreaper.ddns.net/projects",
            "email": "AndrewYushchenko@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/YushchenkoAndrew/API_Server/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:31337",
    "basePath": "/api",
    "paths": {
...
        "/info/range": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Get Info data by date Range",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CreatedAt date \u003e= start",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CreatedAt date \u003c= end",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Column which final result will be sorted by",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Sort by direction",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Info"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/info/sum": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Get Info Sum",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.StatInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Server side error: Something went wrong"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "ERR"
                }
            }
        },
        "models.Info": {
            "type": "object",
            "properties": {
                "Clicks": {
                    "type": "integer",
                    "example": 2
                },
                "Countries": {
                    "type": "string",
                    "example": "UK,US"
                },
                "CreatedAt": {
                    "type": "string",
                    "example": "2021-08-06"
                },
                "ID": {
                    "type": "integer"
                },
                "Media": {
                    "type": "integer",
                    "example": 3
                },
                "Views": {
                    "type": "integer",
                    "example": 1
                },
                "Visitors": {
                    "type": "integer",
                    "example": 4
                }
            }
        },
...
        "models.StatInfo": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer"
                },
                "media": {
                    "type": "integer"
                },
                "views": {
                    "type": "integer"
                },
                "visitors": {
                    "type": "integer"
                }
            }
        },
        "models.Success": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "integer",
                    "example": 1
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "page": {
                    "type": "integer",
                    "example": 0
                },
                "result": {
                    "type": "object"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                },
                "totalItems": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
...
    }
}